+++ src/ui/mainwindow/modeldataholder.cpp
@@ -377,6 +377,14 @@ void ModelDataHolder::setData(QSharedPointer<DataClass> data, QSharedPointer<Cha
         m_modelsWidget->setMetaTab(m_metamodelwidget);
         m_TitleBarWidget->HideModelTools();
     }
+
+    connect(data.data(), &DataClass::Message, data.data(), [this](const QString& str, int level) {
+        QMessageBox::warning(this, tr("Not an exception ... ."), str, QMessageBox::Ok | QMessageBox::Default);
+    });
+
+    connect(data.data(), &DataClass::Warning, data.data(), [this](const QString& str, int level) {
+        QMessageBox::warning(this, tr("Not an exception ... ."), str, QMessageBox::Ok | QMessageBox::Default);
+    });
 }
 
 void ModelDataHolder::SetProjectTabName()